pyproject.toml
[project]
name = "tyroo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "swifter>=1.4.0",
]
